Przebudowa do plików .h + poprawki
